+++ packages/google_maps_flutter/google_maps_flutter_android/example/test/fake_google_maps_flutter_platform.dart
@@ -192,6 +192,11 @@ class FakeGoogleMapsFlutterPlatform extends GoogleMapsFlutterPlatform {
   }
 
   @override
+  Stream<MarkerHoverEvent> onMarkerHover({required int mapId}) {
+    return mapEventStreamController.stream.whereType<MarkerHoverEvent>();
+  }
+
+  @override
   Stream<PolylineTapEvent> onPolylineTap({required int mapId}) {
     return mapEventStreamController.stream.whereType<PolylineTapEvent>();
   }